Permutation

Description

From the book Algorithm Design with Haskell by R. Bird and J. Gibbons.

Documentation

inserts :: a -> [a] -> [[a]]

Insert character into list at each location.

Example:

>>> inserts 'a' "bc"
["abc","bac","bca"]

perms1 :: [a] -> [[a]]

Generate permutations using list comprehensions.

Example:

>>> perms1 "abc"
["abc","bac","bca","acb","cab","cba"]

perms2 :: [a] -> [[a]]

Generate permutations using list comprehensions.

Example:

>>> perms2 "abc"
["abc","acb","bac","bca","cab","cba"]

perms3 :: [a] -> [[a]]

Use fold to generate permutaions.

picks :: [a] -> [(a, [a])]

Pick each member from list return tuple of member and items remaining.

Example:

>>> picks "abc"
[('a',"bc"),('b',"ac"),('c',"ab")]