Random

Description

Exploration of the use of Random inspired by Haskell Design Patterns by Ryan Lemmer

Example

Simulate roll a fair 6-sided dice n times.

With a seed of 111111 then we expect [4,6,5,3,2].

Documentation

dice :: (Int, Int)

Range for a 6-sided dice.

roll :: Int -> IO [Int]

Return n rolls of the dice using the system random generator.

rolls :: StdGen -> [Int]

Produce an infinite stream of dice rolls.

seed :: Int -> StdGen

Seed the dice's random generator.