Trim

Description

An inefficient alternative to dropWhile.

Examples

>>> trim " hello world "
"hello world"

>>> dropWhile' isSpace " hello world "
"hello word "

>>> dropWhileEnd' isSpace " hello world "
" hello word"

Documentation

dropWhile' :: (a -> Bool) -> [a] -> [a]

Drop from beginning of list while predicate is true.

dropWhileEnd' :: (a -> Bool) -> [a] -> [a]

Drop from end of list while predicate is true.

trimStart :: String -> String

Trim white space from start of string.

trimEnd :: String -> String

Trim white space from end of string.

trim :: String -> String

Trim white space from start and end of string.