Weekday

Description

A simple demonstration on how to map to and from a bounded, enumerated data type.

In this example we use the weekdays names as our type to iterate across.

makeWeekday is a simple conversion function to map a Weekday from a string

fullWeek returns a list of all abbreviated weekdays

capitalise returns the title case of a string

The makeWeekday code is rather weak as any word where the first 3 letters map to a weekday is valid for conversion.

Other Examples

Using enum:

This example shows how use Weekday from an enumeration:

>>> toEnum 0 :: WeekDay
Mon

Documentation

data Weekday

Abbreviated days of the week from Mon (Monday) to Sun (Sunday).

Constructors

Mon

Tue

Wed

Thu

Fri

Sat

Sun

makeWeekday :: String -> Maybe Weekday

Read a Weekday from a String.

Some string to Weekday examples:

>>> makeWeekday "tHU"
Just Thu

>>> makeWeekday "sun"
Just Sun

>>> makeWeekday "Bad"
Nothing

Any string bigger than 3 characters also fails: >>> makeWeekday Mond Nothing

fullWeek :: [Weekday]

List all days of the week.

>>> fullWeek
[Mon,Tue,Wed,Thu,Fri,Sat,Sun]

capitalise :: String -> String

Return title case of string.

>>> capitalise "monday"
"Monday"

>>> null (capitalise "")
True