CountEntries

Description

From Chapter 18, Monad Transformers, Real World Haskell by Bryan O'Sullivan, Don Stewart, and John Goerzen.

The function countEntries1 is my version of countEntriesTrad from this chapter (countEntriesS, S means standard or traditional).

The Monad Transformer version is countEntries2 which is my version of countEntries.

GHCi Session

Packages

:m + Control.Monad
:m + System.Directory
:m + System.FilePath

Resources

This is a really good introduction to Monad Transformers, Monday Morning Haskell: Monad Transformers.

The functions countEntriesS and countEntriesU are based off functions from Chapter 18, Monad Transformers, Real World Haskell by Bryan O'Sullivan, Don Stewart, and John Goerzen.

Documentation

countEntries1 :: FilePath -> IO [(FilePath, Int)]

Count entries for a list of paths. (My version.)

Example

What the function returns is a list of tuples of directory and count of entries in that directory:

>>> countEntries1 "public"
[("public",25),("public/src",16)]

Explanation

The function composes a number of different system calls. But the process is simple but clunky using traditional methods. The process will be much simplified once when Monad Transformers are used in countEntries2.

>>> p = "public"

Get contents of path ... and filter to report directories only:

>>> (getDirectoryContents p) >>= filterM (\n -> doesDirectoryExist (p </> n))
["..","src","."]

List directory ignores current and parent directories:

>>> listDirectory p >>= filterM (\n -> doesDirectoryExist (p </> n))
["src"]

Some ways to count number of entries in the path:

>>> ps <- listDirectory p

>>> length ps
25

Same as:

>>> listDirectory p >>= return . length
25

Which is equivalent to:

>>> liftM length (listDirectory p)
25

Recurse into subdirectories:

>>> listDirectory p >>= filterM (\n -> doesDirectoryExist (p </> n)) >>= mapM_ print
"src"

countEntries2 :: FilePath -> WriterT [(FilePath, Int)] IO ()

Count entries in directories for given path. My version using WriterT.

countEntriesS :: FilePath -> IO [(FilePath, Int)]

Count entries in directories for given path. Standard version as documented in Real World Haskell.

>>> p = "public"

>>> :t listDirectory p
listDirectory p :: IO [FilePath]

>>> countEntriesS p
[("public",25),("public/src",16)]

countEntriesU :: FilePath -> WriterT [(FilePath, Int)] IO ()

Count entries in directories for given path. Updated version using WriterT.