Cps

Description

Example from Wikibooks Continuation Passing Style

This uses the Cont Monad from the transformers package.

Using callCC is better than using return as argument is a function and it returns a suspended computation.

Notes

A function of type 'a -> b' would become 'a -> (b -> r) -> r' in CPS, where 'b -> r' is the continuation.

Functions

addCont :: Int -> Int -> Cont r Int

Continuation passing style (CPS) examples.

The following uses the Cont Monad from the transformers package.

Continuation for add function.

pythagorasCont :: Int -> Int -> Cont r Int

Continuation for pythagoras function.

squareCont :: Int -> Cont r Int

Continuation for square function.

toCPS :: a -> forall r. (a -> r) -> r

Custom continuation passing style.

The following example uses a custom continuation passing style (CPS) to demonstrate the general concept.

This converts any value into a suspended computation.

flip ($) :: b -> (b -> c) -> c

Longhand this is: toCPS a = k -> k a

Or more simply as: toCPS a k = k a

fromCPS :: (forall r. (a -> r) -> r) -> a

From custom continuation.

Passing id to the continuation function will return the value.

($ id) :: ((a -> a) -> b) -> b

Longhand this is: fromCPS f = f id