Cps

Description

Example from Wikibooks Continuation Passing Style.

Uses the Cont monad from the transformers package.

Using callCC is better than using return as argument is a function and it returns a suspended computation.

Documentation

addCont :: Int -> Int -> Cont r Int

Continuation for add function.

pythagorasCont :: Int -> Int -> Cont r Int

Continuation for pythagoras function.

squareCont :: Int -> Cont r Int

Continuation for square function.