MyType

Description

Documentation

typeName :: forall a. Typeable a => String

Proxy to record type. Proxy is equivalent to the unit type, (). But also has a phantom type parameter a, which is used to keep track of the type. (Use type from Proxy) data Proxy a = Proxy

Show ambigous types. Hindley-Milner type system can only infer types to the right of the context arrow (=>), which means the type parameter a can never be correctly inferred. Haskell refers to such a type as ambiguous.

>>> typeName @Int
"Int"

>>> typeName @String
"[Char]"

>>> typeName @(Maybe [Int])
"Maybe [Int]"